+++ backend/src/package/add.rs
@@ -48,7 +48,7 @@ pub async fn package_add(
             SystemTime::now()
                 .duration_since(UNIX_EPOCH)
                 .unwrap()
-                .as_secs() as u32,
+                .as_secs() as i64,
         )),
         ..Default::default()
     };
